pyproject.toml
[tool.poetry]
name = "AIvilization"
version = "0.1.0"
description = "Artificial Civilization"
authors = ["Taeho Lee <taeho@corca.ai>", "Chung Hwan Han <hanch@corca.ai>"]
readme = "README.md"

[tool.poetry.scripts]
start = "aivilization.main:main"

[tool.poetry.dependencies]
python = "^3.8"
openai = "^0.27.4"
python-dotenv = "^1.0.0"
pydantic = "^1.10.7"
pinecone-client = "^2.2.1"
retry = "^0.9.2"
asyncio = "^3.4.3"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
ruff = "^0.0.261"
tomli = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# ref: https://beta.ruff.rs/docs/rules/
select = ["E", "F", "UP", "B", "Q"]
ignore = ["UP012", "E722", "E741", "E501", "B007", "UP015"]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.8.
target-version = "py38"

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[tool.ruff.per-file-ignores]
"template.py" = ["E501"]